refactor: 패키지명을 구현하는 내용과 일치하게끔 변경
